change the domain structure by adding shared folder
