--- Domain/Repositories/IRoleRepository.cs
@@ -1,7 +1,0 @@
-﻿
-namespace Domain.Repositories
-{
-    public interface IRoleRepository : IGenericRepository<Role>
-    {
-    }
-}
