+++ Domain/Share/Base/IDomainEvent.cs
@@ -4,10 +4,10 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Domain.Base
+namespace Domain.Share.Base
 {
-    public interface IHasKey<TKey>
+    public interface IDomainEvent : INotification
     {
-        TKey Id { get; set; }
+        DateTime OccurredOn { get; }
     }
 }
