+++ Domain/Share/Events/EntityCreatedEvent.cs
@@ -1,11 +1,6 @@
-﻿using Domain.Base;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using Domain.Share.Base;
 
-namespace Domain.Events
+namespace Domain.Share.Events
 {
     public class EntityCreatedEvent<T> : IDomainEvent where T : BaseEntity<Guid>
     {
