+++ Domain/Share/Events/EntityUpdatedEvent.cs
@@ -1,15 +1,15 @@
-﻿using Domain.Base;
+﻿using Domain.Share.Base;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Domain.Events
+namespace Domain.Share.Events
 {
-    public class EntityDeletedEvent<T> : IDomainEvent where T : BaseEntity<Guid>
+    public class EntityUpdatedEvent<T> : IDomainEvent where T : BaseEntity<Guid>
     {
-        public EntityDeletedEvent(T entity, DateTime eventDateTime)
+        public EntityUpdatedEvent(T entity, DateTime eventDateTime)
         {
             Entity = entity;
             EventDateTime = eventDateTime;
@@ -18,6 +18,7 @@ namespace Domain.Events
         public T Entity { get; }
 
         public DateTime EventDateTime { get; }
+
         public DateTime OccurredOn => throw new NotImplementedException();
     }
 }
