+++ Domain/Share/Repositories/IUnitOfWorkOfT.cs
@@ -1,4 +1,4 @@
-﻿namespace Domain.Repositories
+﻿namespace Domain.Share.Repositories
 {
     public interface IUnitOfWork<TContext> : IUnitOfWork where TContext : DbContext
     {
